+++ src/lib.rs
@@ -24,20 +24,5 @@ pub fn guess_final(word: &str) -> char {
 }
 // 테스트
 pub fn postfix(word: &str, tossi: &str) -> String {
-    let temp = Tossi::new(tossi);
-    println!(
-        "입력된 토시: {:?}, 토시 종류: {:?}",
-        temp.modified, temp.kind
-    );
-
-    let result = match temp.kind {
-        identifier::TossiKind::Neun => particle::neun::change(&word),
-        identifier::TossiKind::Ka => particle::ka::change(&word),
-        identifier::TossiKind::Ro => particle::ro::change(&word),
-        identifier::TossiKind::Ida => particle::ida::change(&word),
-        identifier::TossiKind::None => " ".to_string(),
-    };
-
-    let front = word.to_string();
-    front + &result
+    identifier::postfix(word, tossi)
 }
